modified ParOptTrustRegion methods to optimize for the step, not the design point; added getOptimizedPoint() method to ParOptTrustRegion; modified random_convex/random_quadratic examples to work with complex variables

+++ examples/random_quadratic/random_quadratic.py
@@ -39,7 +39,7 @@ class Quadratic(ParOpt.Problem):
 
         # Evaluate the objective and constraints
         fail = 0
-        con = np.zeros(1)
+        con = np.zeros(1, dtype=ParOpt.dtype)
 
         fobj = 0.5*np.dot(x, np.dot(self.A, x)) + np.dot(self.b, x)
         con[0] = np.dot(x, self.Acon) + self.bcon
